adding wording on resource constrained devices
An open question is about defining a new content type. An alternative
(when using HTTP) would be to use an HTTP Link header to reference the
JSON-LD default context. However, I believe that this would make it
harder to implement the Web of Things Framework on low end IoT devices
which are expected to form the majority of IoT devices. Moreover, HTTP
is a poor fit for such devices due to the increased memory requirements
for handling HTTP compared to lightweight UDP based protocols.