Generalising from devices to things
Many IoT standards define individual capabilities for use in combination. Different organisations use different terms, e.g. modules, resources, capabilities etc. We can use things or partial things. We further allow for collections of things, e.g. a thing as a controller for a set of lights, each of which is a thing.