Experiments with a draft TD spec
This defines the core interaction model directly in terms of Linked Data and without a dependency on JSON Schema. It also includes a proposed serialisation as JSON which is often simpler and easier to understand compared to JSON-LD. An algorithm is defined for mapping the JSON to Linked Data. This has been implemented with a wide range of use cases.