error handling
Web developers may forget or fail to see the importance of declaring the context that defines the mapping from application specific predicates to RDF URIs. This potentially common error could be handled by generating a blank node for the predicate along wit a triple that binds it to the predicate's name as given in the JSON input.