Update privacy.md
This is a great draft.
A couple of comments:

1. Most of these "Mandatory privacy mitigations" only are applicable to devices containing PII.
I understand your goal of minimizing the exposed information, however this requires a communication mechanism about the capabilities / information needs of the consumer. Even if such mechanism would be in place, you can still write a malicious client that just asks "give me everything you can provide".

2. It does not make sense to remove all ids, titles, other data from devices that are not critical, e.g. in home scenarios you may want to clearly identify all Hue lamps and associate them to specific rooms. In the home we make some "trust" assumptions about what devices may expose. However we might want to revisit that and also recommend to assume an "untrusted" relationship and require authentication/authorisation. 

3. I would not communicate outside the WG about the discussion wrt normative/informative architecture spec.
This conversation just started within the WoT group and has not reached consensus.